Move featurecount normalization to module
